// apps/creator-hub/modules/experience-genre/enums/GenreType.ts
enum GenreType {
  Action = 'action',
  OtherAction = 'other_action',
  BattlegroundsAndFighting = 'battlegrounds_and_fighting',
  OpenWorldAction = 'open_world_action',
  MusicAndRhythm = 'music_and_rhythm',
  Adventure = 'adventure',
  OtherAdventure = 'other_adventure',
  Exploration = 'exploration',
  ScavengerHunt = 'scavenger_hunt',
  Story = 'story',
  Education = 'education',
  Entertainment = 'entertainment',
  OtherEntertainment = 'other_entertainment',
  MusicAndAudio = 'music_and_audio',
  ShowcaseAndHub = 'showcase_and_hub',
  Video = 'video',
  ObbyAndPlatformer = 'obby_and_platformer',
  OtherObbyAndPlatformer = 'other_obby_and_platformer',
  ClassicObby = 'classic_obby',
  Runner = 'runner',
  TowerObby = 'tower_obby',
  PartyAndCasual = 'party_and_casual',
  OtherPartyAndCasual = 'other_party_and_casual',
  ChildhoodGame = 'childhood_game',
  ColoringAndDrawing = 'coloring_and_drawing',
  Minigame = 'minigame',
  Quiz = 'quiz',
  Puzzle = 'puzzle',
  OtherPuzzle = 'other_puzzle',
  EscapeRoom = 'escape_room',
  MatchAndMerge = 'match_and_merge',
  Word = 'word',
  Rpg = 'rpg',
  OtherRpg = 'other_rpg',
  ActionRpg = 'action_rpg',
  TurnBasedRpg = 'turn_based_rpg',
  OpenWorldAndSurvivalRpg = 'open_world_and_survival_rpg',
  RoleplayAndAvatarSim = 'roleplay_and_avatar_sim',
  OtherRoleplayAndAvatarSim = 'other_roleplay_and_avatar_sim',
  AnimalSim = 'animal_sim',
  DressUp = 'dress_up',
  Life = 'life',
  MorphRoleplay = 'morph_roleplay',
  PetCare = 'pet_care',
  Shooter = 'shooter',
  OtherShooter = 'other_shooter',
  BattleRoyale = 'battle_royale_shooter',
  DeathmatchShooter = 'deathmatch_shooter',
  PveShooter = 'pve_shooter',
  Shopping = 'shopping',
  OtherShopping = 'other_shopping',
  AvatarShopping = 'avatar_shopping',
  Simulation = 'simulation',
  OtherSimulation = 'other_simulation',
  Idle = 'idle',
  IncrementalSimulator = 'incremental_simulator',
  PhysicsSim = 'physics_sim',
  Sandbox = 'sandbox',
  Tycoon = 'tycoon',
  VehicleSim = 'vehicle_sim',
  Social = 'social',
  SportsAndRacing = 'sports_and_racing',
  OtherSportsAndRacing = 'other_sports_and_racing',
  Racing = 'racing',
  Sports = 'sports',
  Strategy = 'strategy',
  OtherStrategy = 'other_strategy',
  BoardAndCardGames = 'board_and_card_games',
  TowerDefense = 'tower_defense',
  Survival = 'survival',
  OtherSurvival = 'other_survival',
  OneVsAll = 'one_vs_all',
  Escape = 'escape',
  UtilityAndOther = 'utility_and_other',
  NA = 'na',
}

export default GenreType;